tests: delete fuzzer v1
